Create Criador class and move creation and ownership of Dados to it
